renamed o sticky to container
