+++ src/components/o-sticky-container/index.scss
