+++ src/components/o-sticky-container/scss/sticky.scss
